@@ -9,21 +9,38 @@
       }
     },
     "require": {
-        "php": ">=7.0",
-        "jms/serializer": "~1.0|~2.0",
-        "symfony/console": "~2.8|~3.0",
-        "symfony/framework-bundle": "~2.8|~3.0",
-        "symfony/property-access": "~2.8|~3.0",
-        "symfony/expression-language": "~2.8|~3.0"
+        "php": ">=7.4 || ^8.0",
+        "ext-simplexml": "*",
+        "ext-dom": "*",
+        "symfony/console": "~2.8 || ~3.0 || ^4.4",
+        "symfony/framework-bundle": "~2.8 || ~3.0 || ^4.4",
+        "symfony/property-access": "~2.8 || ~3.0 || ^4.4",
+        "symfony/expression-language": "~2.8 || ~3.0 || ^4.4",
+        "jms/serializer": "^3.0",
+        "symfony/dotenv": "~2.8 || ~3.0"
     },
     "require-dev": {
-        "jms/serializer-bundle": "^2.0.0",
-        "predis/predis": "^1.0",
-        "snc/redis-bundle": "~2.0",
-        "symfony/monolog-bundle": "~2.4|~3.0",
+        "roave/security-advisories": "dev-latest",
+        "predis/predis": "dev-main",
+        "snc/redis-bundle": "^3.6",
+        "symfony/monolog-bundle": "~2.4 || ~3.0 || ^4.4",
         "symfony/phpunit-bridge": "^4.2",
-        "symfony/validator": "~2.8|~3.0"
+        "symfony/validator": "~2.8 || ~3.0 || ^4.4",
+        "jms/serializer-bundle": "^3.10",
+        "phpunit/phpunit": "^9.5",
+        "symfony/translation": "~2.8 || ~3.0 || ^4.4 || ^5.4"
     },
     "minimum-stability": "dev",
-    "prefer-stable" : true
+    "prefer-stable" : true,
+    "config": {
+        "allow-plugins": {
+            "symfony/flex": true
+        }
+    },
+    "scripts": {
+        "auto-scripts": {
+            "cache:clear": "symfony-cmd",
+            "assets:install %PUBLIC_DIR%": "symfony-cmd"
+        }
+    }
 }
