@@ -14,12 +14,12 @@ trait HasAPIVersion
      *
      * @var string
      */
-    protected $version;
+    protected string $version;
 
     /**
      * @return string
      */
-    public function getAPIVersion()
+    public function getAPIVersion(): string
     {
         return $this->version;
     }
@@ -27,7 +27,7 @@ trait HasAPIVersion
     /**
      * @param string $version
      */
-    public function setAPIVersion($version)
+    public function setAPIVersion(string $version): void
     {
         if (!empty($version) && !\is_string($version)) {
             throw new \InvalidArgumentException('Expected null or string in method setVersion');
