@@ -3,6 +3,7 @@
 namespace Smartbox\CoreBundle\Tests\Utils\Monolog\Formatter;
 
 use JMS\Serializer\SerializerInterface;
+use Smartbox\CoreBundle\Tests\AppKernel;
 use Smartbox\CoreBundle\Tests\Fixtures\Entity\TestEntity;
 use Smartbox\CoreBundle\Utils\Monolog\Formatter\JMSSerializerFormatter;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
@@ -19,17 +20,17 @@ class JMSSerializerFormatterTest extends WebTestCase
 
     public static function getKernelClass()
     {
-        return \Smartbox\CoreBundle\Tests\AppKernel::class;
+        return AppKernel::class;
     }
 
-    protected function setUp()
+    protected function setUp(): void
     {
         static::bootKernel();
         $container = static::$kernel->getContainer();
         $this->serializer = $container->get('jms_serializer');
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         parent::tearDown();
         self::$class = null;
