@@ -35,7 +35,7 @@ class JMSSerializerFormatter extends JsonFormatter
     /**
      * {@inheritdoc}
      */
-    public function format(array $record)
+    public function format(array $record): string
     {
         $context = new SerializationContext();
         $context->setGroups(['logs']);
